requirements.txt
streamlit
numpy
pillow
onnxruntime
